// components/Profile/types.ts
export interface Profile {
    id: number;
    bio: string;
    avatar: string;
    id_number: string;
    address: string;
    city: string;
    emergency_contact_name: string;
    emergency_contact_phone: string;
    emergency_contact_relationship: string;
    prefers_email: boolean;
    prefers_sms: boolean;
    prefers_phone: boolean;
    user: string;
}